Revert "Revert "fix memory access issue on raspbi""
This reverts commit 389ac8aa6c826e5cd5adfc240820961c0046d24f.

+++ backend/finance/domain/event.go
@@ -7,7 +7,7 @@ type TransactionRequested struct {
 	id          string
 	recipientId string
 	senderId    string
-	amount      uint32
+	amount      uint64
 }
 
 func (t TransactionRequested) Id() string {
@@ -22,11 +22,11 @@ func (t TransactionRequested) SenderId() string {
 	return t.senderId
 }
 
-func (t TransactionRequested) Amount() uint32 {
+func (t TransactionRequested) Amount() uint64 {
 	return t.amount
 }
 
-func NewTransactionRequest(id string, recipientId string, senderId string, amount uint32) TransactionRequested {
+func NewTransactionRequest(id string, recipientId string, senderId string, amount uint64) TransactionRequested {
 
 	if amount <= 0 {
 		panic("amount has to be greater than 0")
